Canonicalize predicates using Equivalences
